tools/legal-doc-creator/requirements.txt
markdown>=3.5.0
python-docx>=1.1.0
